pyproject.toml
[tool.poetry]
name = "tg-bot"
version = "0.1.0"
description = "Telegram bot with SQLite database"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiogram = "^2.25"
aiosqlite = "^0.19.0"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
black = "^23.3.0" 